+++ packages/@runlightyear/airtable/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @runlightyear/airtable
 
+## 0.7.1
+
+### Patch Changes
+
+- Updated dependencies [efed4a0c7]
+  - @runlightyear/lightyear@0.36.1
+
 ## 0.7.0
 
 ### Minor Changes