+++ packages/@runlightyear/cli/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @runlightyear/cli
 
+## 0.28.1
+
+### Patch Changes
+
+- Updated dependencies [efed4a0c7]
+  - @runlightyear/lightyear@0.36.1
+
 ## 0.28.0
 
 ### Minor Changes